requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt config/requirements.in
#
amqp==5.0.9
    # via kombu
asgiref==3.5.0
    # via django
billiard==3.6.4.0
    # via celery
cachetools==5.0.0
    # via google-auth
celery==5.2.3
    # via django-celery-beat
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
django==3.2.12
    # via
    #   django-celery-beat
    #   django-timezone-field
django-celery-beat==2.2.1
    # via -r config/requirements.in
django-timezone-field==4.2.3
    # via django-celery-beat
google-api-core==2.5.0
    # via google-api-python-client
google-api-python-client==2.36.0
    # via pygsheets
google-auth==2.6.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.4.6
    # via pygsheets
googleapis-common-protos==1.54.0
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
kombu==5.2.3
    # via celery
oauthlib==3.2.0
    # via requests-oauthlib
prompt-toolkit==3.0.26
    # via click-repl
protobuf==3.19.4
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygsheets==2.0.5
    # via -r config/requirements.in
pyparsing==3.0.7
    # via httplib2
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python-decouple==3.6
    # via -r config/requirements.in
pytz==2021.3
    # via
    #   celery
    #   django
    #   django-timezone-field
requests==2.27.1
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
sentry-sdk==1.5.4
    # via -r config/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.8
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
